requirements.txt
aiofiles==0.6.0
aiohttp==3.7.2
allennlp==1.2.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
blis==0.7.3
boto3==1.16.23
botocore==1.19.23
bs4==0.0.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cymem==2.0.4
dataclasses==0.6
DAWG-Python==0.7.2
decorator==4.4.2
docopt==0.6.2
fastapi==0.61.2
filelock==3.0.12
funcy==1.15
future==0.18.2
gensim==3.8.3
google==3.0.0
googleapis-common-protos==1.6.0
grpcio==1.33.2
grpcio-tools==1.33.2
h11==0.11.0
h5py==3.1.0
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==3.1.0
iniconfig==1.1.1
instagram-scraper==1.9.1
install==1.3.4
intervaltree==3.1.0
ipymarkup==0.9.0
isort==5.6.4
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonnet==0.17.0
jsonpickle==1.4.1
jsonschema==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
moviepy==1.0.3
multidict==5.0.2
multipart==0.2.3
murmurhash==1.0.4
natasha==1.4.0
navec==0.10.0
networkx==2.5
nltk==3.5
nose==1.3.7
numexpr==2.7.1
numpy==1.19.4
opuslib==3.0.1
overrides==3.1.0
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
pipreqs==0.4.10
plac==1.1.3
pluggy==0.13.1
preshed==3.0.4
proglog==0.1.9
protobuf==3.12.4
psycopg2-binary==2.8.6
py==1.9.0
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.7.2
pydub==0.24.1
pyLDAvis==2.1.2
pylint==2.6.0
pymorphy2==0.9.1
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pyrsistent==0.17.3
pyTelegramBotAPI==3.7.4
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2020.4
razdel==0.5.0
regex==2020.11.13
requests==2.22.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
schedule==0.6.0
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.91
six==1.12.0
sklearn==0.0
slovnet==0.5.0
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.0.1
spacy==2.3.3
srsly==1.0.4
starlette==0.13.6
Telethon==1.17.5
tensorboardX==2.1
textsplit==0.5
thinc==7.4.3
threadpoolctl==2.1.0
tinkoff-voicekit-client==0.2.2
tokenizers==0.9.3
toml==0.10.2
torch==1.7.0
torchaudio==0.7.0
torchtext==0.8.0
tqdm==4.51.0
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.12.2
wasabi==0.8.0
wrapt==1.12.1
yarg==0.1.9
yargy==0.15.0
yarl==1.6.3
zipp==3.4.0
